refactor(claude_statusline): restructure models and improve rendering layout
- Relocated monolithic `models.py` into a modular `types/` directory containing `payload.py`, `layout.py`, `git.py`, and `cache.py`.
- Updated all codebase and test imports appropriately.
- Overhauled rendering strategy in `render.py` to utilize `Table.grid()` per individual line instead of a single shared table, fixing column width sync misalignments across independent lines.
- Factored in the terminal width limit using `shutil.get_terminal_size()` to prevent unintended console wrapping overflows.
