+++ src/python/claude_statusline/claude_statusline/render.py
@@ -1,11 +1,14 @@
+import shutil
 from collections.abc import Iterable
 
 from rich.console import Console
 from rich.table import Table
 from rich.text import Text
 
-from claude_statusline.models import GitInfo, SegmentGenerationResult, StatusLineStdIn
 from claude_statusline.segments.constants import get_icon
+from claude_statusline.types.git import GitInfo
+from claude_statusline.types.layout import SegmentGenerationResult
+from claude_statusline.types.payload import StatusLineStdIn
 
 
 def render_lines(
@@ -21,32 +24,36 @@ def render_lines(
     if not segments_list:
         return []
 
-    # Find unique columns (indices) and rows (lines)
-    all_indices = sorted({seg.index for seg in segments_list})
+    term_size = shutil.get_terminal_size((80, 24))
+    # We leave a buffer of ~20 columns to account for Claude's own trailing UI.
+    # Claude uses some space on the right, and maybe a few chars for borders.
+    safe_width = max(20, term_size.columns - 20)
+
+    # Find unique rows (lines)
     all_lines = sorted({seg.line for seg in segments_list})
 
     # Group segments by line then by index
-    # We might have multiple segments with the same line and index, though
-    # ideally we wouldn't. We'll join them with sep if they exist.
     grid_data = {}
     for line in all_lines:
         grid_data[line] = {}
-        for index in all_indices:
-            grid_data[line][index] = []
 
     for seg in segments_list:
+        if seg.index not in grid_data[seg.line]:
+            grid_data[seg.line][seg.index] = []
         grid_data[seg.line][seg.index].append(seg)
 
-    # Build the rich table
-    table = Table(show_header=False, box=None, pad_edge=False, padding=(0, 2))
+    output_lines = []
 
-    # Add columns for each unique index
-    for _ in all_indices:
-        table.add_column()
-
+    # Build a separate table per line to avoid columns syncing widths across different lines
     for line in all_lines:
+        table = Table.grid(padding=(0, 2))
+
+        indices = sorted(grid_data[line].keys())
+        for _ in indices:
+            table.add_column(justify="left", no_wrap=True)
+
         row_cells = []
-        for index in all_indices:
+        for index in indices:
             segs = grid_data[line][index]
             # Multiple segments in the same cell get joined
             text_str = sep.join(s.segment.text for s in sorted(segs, key=lambda x: x.generator))
@@ -56,14 +63,16 @@ def render_lines(
             else:
                 row_cells.append(Text.from_ansi(""))
 
-        table.add_row(*row_cells)
+        if row_cells:
+            table.add_row(*row_cells)
 
-    # Use a wide console to prevent line wrapping during string extraction
-    console = Console(width=2000, force_terminal=True, color_system="truecolor", highlight=False)
-    with console.capture() as capture:
-        console.print(table)
+            # Use a bounded console to allow rich to gracefully truncate or handle the line
+            console = Console(width=safe_width, force_terminal=True, color_system="truecolor", highlight=False)
+            with console.capture() as capture:
+                console.print(table)
 
-    output_lines = capture.get().splitlines()
+            captured_lines = capture.get().splitlines()
+            output_lines.extend(captured_lines)
 
     # Strip any extra completely empty lines at the end that Rich might add,
     # but keep lines with content (even just spaces if they were formatted)
