+++ src/python/claude_statusline/claude_statusline/types/git.py
@@ -1,0 +1,14 @@
+from pydantic import BaseModel
+
+
+class GitInfo(BaseModel):
+    branch: str
+    remote: str | None
+    dirty: bool
+    staged: bool
+    untracked: bool
+    ahead: int
+    behind: int
+    is_repo: bool
+    is_worktree: bool = False
+    stash_count: int = 0
