+++ src/python/claude_statusline/tests/test_main.py
@@ -9,8 +9,8 @@ import pytest
 from typer.testing import CliRunner
 
 from claude_statusline.main import cli, run_external_generator
-from claude_statusline.models import Segment, SegmentGenerationResult
 from claude_statusline.segments.workspace import shorten_path
+from claude_statusline.types.layout import Segment, SegmentGenerationResult
 
 
 class TestMainSmoke(unittest.TestCase):
